Create clientSync.js
Created the JavaScript code that should always be running on the client as a node.js service. This attempts connections to the server and sync files between the server and itself. Can be running on multiple clients with the same credentials allowing for many computers to sync file between each other. Requires two settings files in AppData\FileSync\